requirements.txt
openai
pypdf2 # For PDF text extraction
python-magic # For basic file type detection 
